feat/AB#104296_EIS-add-date-field-calc-method-for-html-widgets refactor: helper functions for html parser and grid data formatter to standalone classes in order to be able to use injected pipes feat: apply changes to related components/files using aforementioned helpers
