+++ libs/shared/src/index.ts
@@ -3,112 +3,114 @@
  */
 
 // === SERVICES ===
-export * from './lib/services/auth/auth.service';
-export * from './lib/services/grid/grid.service';
-export * from './lib/services/application/application.service';
-export * from './lib/services/download/download.service';
-export * from './lib/services/workflow/workflow.service';
-export * from './lib/services/form-builder/form-builder.service';
-export * from './lib/services/form/form.service';
-export * from './lib/services/api-proxy/api-proxy.service';
-export * from './lib/services/dashboard/dashboard.service';
-export * from './lib/services/grid-layout/grid-layout.service';
-export * from './lib/services/reference-data/reference-data.service';
-export * from './lib/services/kendo-translation/kendo-translation.service';
-export * from './lib/services/date-translate/date-translate.service';
-export * from './lib/services/breadcrumb/breadcrumb.service';
 export * from './lib/services/aggregation/aggregation.service';
+export * from './lib/services/api-proxy/api-proxy.service';
+export * from './lib/services/application/application.service';
 export * from './lib/services/auth-interceptor/auth-interceptor.service';
+export * from './lib/services/auth/auth.service';
+export * from './lib/services/breadcrumb/breadcrumb.service';
 export * from './lib/services/confirm/confirm.service';
 export * from './lib/services/context/context.service';
-export * from './lib/services/data-template/data-template.service';
-export * from './lib/services/editor/editor.service';
-export * from './lib/services/rest/rest.service';
-export * from './lib/services/map/map-layers.service';
 export * from './lib/services/dashboard-automation/dashboard-automation.service';
+export * from './lib/services/dashboard/dashboard.service';
+export * from './lib/services/data-template/data-template.service';
+export * from './lib/services/date-translate/date-translate.service';
+export * from './lib/services/download/download.service';
+export * from './lib/services/editor/editor.service';
+export * from './lib/services/form-builder/form-builder.service';
+export * from './lib/services/form/form.service';
+export * from './lib/services/grid-data-formatter/grid-data-formatter.service';
+export * from './lib/services/grid-layout/grid-layout.service';
+export * from './lib/services/grid/grid.service';
+export * from './lib/services/html-parser/html-parser.service';
+export * from './lib/services/kendo-translation/kendo-translation.service';
+export * from './lib/services/map/map-layers.service';
+export * from './lib/services/reference-data/reference-data.service';
+export * from './lib/services/rest/rest.service';
+export * from './lib/services/workflow/workflow.service';
 
 // === DIRECTIVES ===
-export * from './lib/directives/skeleton/public-api';
-export * from './lib/directives/fullscreen/public-api';
 export * from './lib/directives/async-monaco-editor/public-api';
+export * from './lib/directives/fullscreen/public-api';
+export * from './lib/directives/skeleton/public-api';
 
 // === GUARDS ===
 export * from './lib/guards/permission.guard';
 
 // === MODELS ===
+export * from './lib/models/aggregation.model';
+export * from './lib/models/api-configuration.model';
+export * from './lib/models/application.model';
+export * from './lib/models/automation.model';
+export * from './lib/models/channel.model';
 export * from './lib/models/dashboard.model';
 export * from './lib/models/form.model';
-export * from './lib/models/record.model';
-export * from './lib/models/resource.model';
-export * from './lib/models/user.model';
-export * from './lib/models/application.model';
-export * from './lib/models/page.model';
-export * from './lib/models/workflow.model';
-export * from './lib/models/step.model';
+export * from './lib/models/layout.model';
+export * from './lib/models/metadata.model';
 export * from './lib/models/notification.model';
-export * from './lib/models/channel.model';
-export * from './lib/models/subscription.model';
+export * from './lib/models/page.model';
 export * from './lib/models/position-attribute-category.model';
 export * from './lib/models/position-attribute.model';
-export * from './lib/models/api-configuration.model';
 export * from './lib/models/pullJob.model';
-export * from './lib/models/layout.model';
-export * from './lib/models/aggregation.model';
+export * from './lib/models/record.model';
 export * from './lib/models/reference-data.model';
-export * from './lib/models/metadata.model';
-export * from './lib/models/automation.model';
+export * from './lib/models/resource.model';
+export * from './lib/models/step.model';
+export * from './lib/models/subscription.model';
+export * from './lib/models/user.model';
+export * from './lib/models/workflow.model';
 
 // === COMPONENTS ===
-export * from './lib/components/aggregation/edit-aggregation-modal/edit-aggregation-modal.component';
-export * from './lib/components/layout/public-api';
-export * from './lib/components/navbar/public-api';
 export * from './lib/components/access/public-api';
-export * from './lib/components/form/public-api';
+export * from './lib/components/aggregation/edit-aggregation-modal/edit-aggregation-modal.component';
+export * from './lib/components/applications-summary/public-api';
+export * from './lib/components/confirm-modal/public-api';
+export * from './lib/components/content-choice/public-api';
+export * from './lib/components/convert-modal/public-api';
+export * from './lib/components/dashboard-filter-icon/public-api';
+export * from './lib/components/dashboard-filter/public-api';
+export * from './lib/components/dashboard/public-api';
+export * from './lib/components/edit-calculated-field-modal/public-api';
+export * from './lib/components/editable-text/public-api';
 export * from './lib/components/form-builder/public-api';
+export * from './lib/components/form/public-api';
+export * from './lib/components/layout/public-api';
+export * from './lib/components/mapping/public-api';
+export * from './lib/components/navbar/public-api';
+export * from './lib/components/payload-modal/payload-modal.component';
+export * from './lib/components/record-history/public-api';
+export * from './lib/components/record-modal/public-api';
+export * from './lib/components/role-summary/public-api';
+export * from './lib/components/roles/public-api';
+export * from './lib/components/search-menu/public-api';
+export * from './lib/components/search-resource-grid-modal/public-api';
+export * from './lib/components/skeleton/skeleton-table/public-api';
+export * from './lib/components/snackbar-spinner/public-api';
+export * from './lib/components/templates/public-api';
+export * from './lib/components/ui/empty/public-api';
+export * from './lib/components/user-summary/public-api';
+export * from './lib/components/users/public-api';
+export * from './lib/components/utils/unsubscribe/public-api';
+export * from './lib/components/widget-grid/public-api';
 export * from './lib/components/widgets/chart-settings/public-api';
 export * from './lib/components/widgets/editor-settings/public-api';
 export * from './lib/components/widgets/grid-settings/public-api';
 export * from './lib/components/widgets/map-settings/public-api';
-export * from './lib/components/widgets/summary-card-settings/public-api';
-export * from './lib/components/widget-grid/public-api';
-export * from './lib/components/confirm-modal/public-api';
-export * from './lib/components/user-summary/public-api';
-export * from './lib/components/users/public-api';
-export * from './lib/components/templates/public-api';
-export * from './lib/components/roles/public-api';
-export * from './lib/components/convert-modal/public-api';
-export * from './lib/components/record-history/public-api';
-export * from './lib/components/record-modal/public-api';
-export * from './lib/components/search-menu/public-api';
-export * from './lib/components/search-resource-grid-modal/public-api';
-export * from './lib/components/content-choice/public-api';
-export * from './lib/components/workflow-stepper/public-api';
-export * from './lib/components/applications-summary/public-api';
-export * from './lib/components/dashboard-filter/public-api';
-export * from './lib/components/dashboard-filter-icon/public-api';
-export * from './lib/components/snackbar-spinner/public-api';
-export * from './lib/components/skeleton/skeleton-table/public-api';
-export * from './lib/components/mapping/public-api';
-export * from './lib/components/role-summary/public-api';
-export * from './lib/components/ui/empty/public-api';
-export * from './lib/components/edit-calculated-field-modal/public-api';
-export * from './lib/components/utils/unsubscribe/public-api';
-export * from './lib/components/editable-text/public-api';
-export * from './lib/components/users/public-api';
-export * from './lib/components/payload-modal/payload-modal.component';
 export * from './lib/components/widgets/map/public-api';
-export * from './lib/components/dashboard/public-api';
+export * from './lib/components/widgets/summary-card-settings/public-api';
+export * from './lib/components/workflow-stepper/public-api';
 
 // Export of controls
-export * from './lib/components/controls/public-api';
 export * from './lib/components/applications-archive/public-api';
 export * from './lib/components/button-action/public-api';
+export * from './lib/components/controls/public-api';
 export * from './lib/components/custom-widget-style/custom-widget-style.component';
 export * from './lib/components/list-filter/list-filter.component';
 export * from './lib/components/status-options/status-options.component';
 
 // === CUSTOM COMPONENTS FOR SURVEY CREATOR PROPERTY GRID === //
 export * from './lib/survey/components/application-dropdown/application-dropdown.component';
+export * from './lib/survey/components/code-editor/code-editor.component';
 export * from './lib/survey/components/geofields-listbox/geofields-listbox.component';
 export * from './lib/survey/components/reference-data-dropdown/reference-data-dropdown.component';
 export * from './lib/survey/components/resource-available-fields/resource-available-fields.component';
@@ -116,7 +118,6 @@ export * from './lib/survey/components/resource-custom-filters/resource-custom-f
 export * from './lib/survey/components/resource-dropdown/resource-dropdown.component';
 export * from './lib/survey/components/resource-select-text/resource-select-text.component';
 export * from './lib/survey/components/test-service-dropdown/test-service-dropdown.component';
-export * from './lib/survey/components/code-editor/code-editor.component';
 
 /** Grid Layouts */
 export * from './lib/components/grid-layout/edit-layout-modal/public-api';
@@ -129,13 +130,13 @@ export * from './lib/components/ui/tagbox/public-api';
 export * from './lib/views/public-api';
 
 // === MODULE ===
-export * from './lib/shared.module';
 export * from './lib/components/widget-choice/widget-choice.module';
+export * from './lib/shared.module';
 
 // === PIPES ===
+export * from './lib/pipes/cron-parser/public-api';
 export * from './lib/pipes/date/public-api';
 export * from './lib/pipes/readable-cron/public-api';
-export * from './lib/pipes/cron-parser/public-api';
 
 // === UTILS ===
 export * from './lib/utils/public-api';
