+++ libs/shared/src/lib/components/query-builder/tab-fields/tab-fields.component.ts
@@ -12,11 +12,11 @@ import {
   ViewContainerRef,
 } from '@angular/core';
 import { UntypedFormArray, UntypedFormGroup } from '@angular/forms';
+import { INLINE_EDITOR_CONFIG } from '../../../const/tinymce.const';
 import { EditorService } from '../../../services/editor/editor.service';
-import { getCalcKeys, getDataKeys } from '../../../utils/parser/utils';
+import { HtmlParserService } from '../../../services/html-parser/html-parser.service';
 import { addNewField } from '../query-builder-forms';
 import { QueryBuilderComponent } from '../query-builder.component';
-import { INLINE_EDITOR_CONFIG } from '../../../const/tinymce.const';
 
 /**
  * Component used for the selection of fields to display the fields in tabs
@@ -55,8 +55,12 @@ export class TabFieldsComponent implements OnInit, OnChanges {
    * Component used for the selection of fields to display the fields in tabs.
    *
    * @param editorService Editor service used to get main URL and current language
+   * @param htmlParserService Html parser service to parse the values for html layout
    */
-  constructor(private editorService: EditorService) {
+  constructor(
+    private editorService: EditorService,
+    private htmlParserService: HtmlParserService
+  ) {
     // Set the editor base url based on the environment file
     this.editor.base_url = editorService.url;
     // Set the editor language
@@ -192,10 +196,10 @@ export class TabFieldsComponent implements OnInit, OnChanges {
     this.fieldForm = this.form.at(index) as UntypedFormGroup;
     if (this.fieldForm.value.kind === 'SCALAR') {
       // Setup field format editor auto completer
-      const dataKeys = getDataKeys([
+      const dataKeys = this.htmlParserService.getDataKeys([
         { name: this.fieldForm.controls.name.value },
       ]);
-      const calcKeys = getCalcKeys();
+      const calcKeys = this.htmlParserService.getCalcKeys();
       const keys = dataKeys.concat(calcKeys);
 
       this.editorService.addCalcAndKeysAutoCompleter(this.editor, keys);
