+++ libs/shared/src/lib/components/ui/core-grid/grid/grid.component.ts
@@ -1,3 +1,5 @@
+import { Dialog } from '@angular/cdk/dialog';
+import { DOCUMENT } from '@angular/common';
 import {
   AfterViewInit,
   Component,
@@ -17,15 +19,33 @@ import {
   ViewChild,
   ViewChildren,
 } from '@angular/core';
+import { UntypedFormControl, UntypedFormGroup } from '@angular/forms';
+import { TranslateService } from '@ngx-translate/core';
+import { SnackbarService, TooltipDirective } from '@oort-front/ui';
+import { ResizeBatchService } from '@progress/kendo-angular-common';
 import {
-  GridComponent as KendoGridComponent,
+  ColumnComponent,
   GridDataResult,
+  GridComponent as KendoGridComponent,
   PageChangeEvent,
   RowArgs,
   SelectionEvent,
-  ColumnComponent,
 } from '@progress/kendo-angular-grid';
-import { Dialog } from '@angular/cdk/dialog';
+import { PopupRef, PopupService } from '@progress/kendo-angular-popup';
+import {
+  CompositeFilterDescriptor,
+  SortDescriptor,
+} from '@progress/kendo-data-query';
+import { get, has, intersection, isEqual, isNil } from 'lodash';
+import { debounceTime, distinctUntilChanged, takeUntil } from 'rxjs/operators';
+import { DownloadService } from '../../../../services/download/download.service';
+import { GridDataFormatterService } from '../../../../services/grid-data-formatter/grid-data-formatter.service';
+import { GridService } from '../../../../services/grid/grid.service';
+import { ResizeObservable } from '../../../../utils/rxjs/resize-observable.util';
+import { UnsubscribeComponent } from '../../../utils/unsubscribe/unsubscribe.component';
+import { WidgetComponent } from '../../../widget/widget.component';
+import { GridLayout } from '../models/grid-layout.model';
+import { GridActions } from '../models/grid-settings.model';
 import {
   EXPORT_SETTINGS,
   GRADIENT_SETTINGS,
@@ -33,27 +53,6 @@ import {
   PAGER_SETTINGS,
   SELECTABLE_SETTINGS,
 } from './grid.constants';
-import {
-  CompositeFilterDescriptor,
-  SortDescriptor,
-} from '@progress/kendo-data-query';
-import { ResizeBatchService } from '@progress/kendo-angular-common';
-import { PopupRef, PopupService } from '@progress/kendo-angular-popup';
-import { UntypedFormControl, UntypedFormGroup } from '@angular/forms';
-import { GridService } from '../../../../services/grid/grid.service';
-import { DownloadService } from '../../../../services/download/download.service';
-import { debounceTime, distinctUntilChanged, takeUntil } from 'rxjs/operators';
-import { GridLayout } from '../models/grid-layout.model';
-import { get, intersection, isNil, has, isEqual } from 'lodash';
-import { TranslateService } from '@ngx-translate/core';
-import { SnackbarService, TooltipDirective } from '@oort-front/ui';
-import { UnsubscribeComponent } from '../../../utils/unsubscribe/unsubscribe.component';
-import { DOCUMENT } from '@angular/common';
-import { WidgetComponent } from '../../../widget/widget.component';
-import { DatePipe } from '../../../../pipes/date/date.pipe';
-import { ResizeObservable } from '../../../../utils/rxjs/resize-observable.util';
-import { formatGridRowData } from './utils/grid-data-formatter';
-import { GridActions } from '../models/grid-settings.model';
 
 /** Minimum column width */
 const MIN_COLUMN_WIDTH = 100;
@@ -75,7 +74,7 @@ const matches = (el: any, selector: any) =>
   selector: 'shared-grid',
   templateUrl: './grid.component.html',
   styleUrls: ['./grid.component.scss'],
-  providers: [PopupService, ResizeBatchService, DatePipe],
+  providers: [PopupService, ResizeBatchService],
 })
 export class GridComponent
   extends UnsubscribeComponent
@@ -293,7 +292,6 @@ export class GridComponent
    *
    * @param widgetComponent parent widget component ( optional )
    * @param environment Current environment
-   * @param datePipe Shared date pipe
    * @param dialog The Dialog service
    * @param gridService The grid service
    * @param renderer The renderer library
@@ -303,11 +301,11 @@ export class GridComponent
    * @param el Ref to html element
    * @param document document
    * @param popupService Kendo popup service
+   * @param gridDataFormatterService GridDataFormatterService
    */
   constructor(
     @Optional() public widgetComponent: WidgetComponent,
     @Inject('environment') environment: any,
-    @Inject(DatePipe) private datePipe: DatePipe,
     private dialog: Dialog,
     private gridService: GridService,
     private renderer: Renderer2,
@@ -316,7 +314,8 @@ export class GridComponent
     private snackBar: SnackbarService,
     private el: ElementRef,
     @Inject(DOCUMENT) private document: Document,
-    private popupService: PopupService
+    private popupService: PopupService,
+    private gridDataFormatterService: GridDataFormatterService
   ) {
     super();
     this.environment = environment.module || 'frontoffice';
@@ -369,7 +368,7 @@ export class GridComponent
       (changes['fields']?.currentValue?.length || this.fields.length)
     ) {
       this.data.data.forEach((gridRow) => {
-        formatGridRowData(gridRow, this.fields, this.datePipe);
+        this.gridDataFormatterService.formatGridRowData(gridRow, this.fields);
       });
     }
     // First load of records, or on page change
@@ -1048,8 +1047,11 @@ export class GridComponent
             case 'datetime-local':
             case 'datetime':
             case 'date': {
-              contentSize = (this.datePipe.transform(data[type.field]) || '')
-                .length;
+              contentSize = (
+                this.gridDataFormatterService.datePipe.transform(
+                  data[type.field]
+                ) || ''
+              ).length;
               break;
             }
             case 'file': {
