+++ libs/shared/src/lib/components/ui/core-grid/grid/grid.module.ts
@@ -13,7 +13,6 @@ import { DateInputsModule } from '@progress/kendo-angular-dateinputs';
 import { DropDownsModule } from '@progress/kendo-angular-dropdowns';
 import { GridModule as KendoGridModule } from '@progress/kendo-angular-grid';
 import { InputsModule } from '@progress/kendo-angular-inputs';
-import { DateModule } from '../../../../pipes/date/date.module';
 import { DateFilterMenuModule } from '../date-filter-menu/date-filter-menu.module';
 import { ExpandedCommentModule } from '../expanded-comment/expanded-comment.module';
 import { ExportModule } from '../export/export.module';
@@ -55,7 +54,6 @@ import { GridFilterMenuModule } from '../filter-menu/filter-menu.module';
     ExportModule,
     // === TRANSLATE ===
     TranslateModule,
-    DateModule,
     TextareaModule,
     uiButtonModule,
   ],
